requirements.txt
APScheduler==3.6.3
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-telegram-bot==13.7
pytz==2021.3
six==1.16.0
toml==0.10.2
tornado==6.1
transitions==0.8.10
tzlocal==3.0
